install/requirements.txt
crcmod==1.7
et-xmlfile==1.1.0
future==0.18.3
iso8601==2.1.0
numpy==1.26.4
openpyxl==3.1.2
paho-mqtt==2.1.0
pandas==2.2.3
pymodbus==3.7.2
pyserial==3.5
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
serial==0.0.97
six==1.16.0
tornado==6.4.1
tzdata==2023.4
XlsxWriter==3.1.9
